requirements.txt
twikit==2.3.3
uvicorn==0.34.0
crewai==0.102.0
langchain-openai==0.3.7
langchain_anthropic==0.3.9
langchain_google_genai==2.0.11
tweepy==4.14.0
python-dotenv==1.0.0
streamlit==1.32.0